+++ CHANGES.md
@@ -48,6 +48,7 @@
 -   pat date picker: Support updating a date if it is before another dependent date.
 -   pat tabs: Refactor based on ``ResizeObserver`` and fix problems calculating the with with transitions.
 -   pat tabs: When clicking on the ``extra-tabs`` element, toggle between ``open`` and ``closed`` classes to allow opening/closing an extra-tabs menu via CSS.
+-   pat autofocus: Do not autofocus in iframes. Fixes: #761.
 
 ### Technical
 
@@ -82,6 +83,9 @@
 -   pat scroll: Fix scrolling offset incorrectly applied. Fixes: #763.
 -   Core registry: Fix ``transformPattern`` to also work with patterns which extend from Base.
     Fixes a problem with pat-auto-suggest not auto submitting.
+-   pat autofocus: Implement documented behavior to not focus on prefilled element, if there is another autofocus element which is empty.
+-   pat autofocus: Instead of calling autofocus for each element call it only once.
+-   pat autofocus: Register event handler only once.
 
 
 ## 3.0.0-dev - unreleased